Removes treatment access level name function
